third_party/WebKit/LayoutTests/fast/dom/ping-attribute-dom-binding.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>

description('Tests the DOM bindings for the ping attribute');

var anchor = document.createElement('a');
shouldBeEqualToString('anchor.ping', '');
anchor.setAttribute('ping', 'p1');
shouldBeEqualToString('anchor.ping', 'p1');
anchor.ping = 'p2';
shouldBeEqualToString('anchor.getAttribute("ping")', 'p2');
anchor.ping = null;
shouldBeEqualToString('anchor.getAttribute("ping")', 'null');
shouldBeEqualToString('anchor.ping', 'null');

var area = document.createElement('area');
shouldBeEqualToString('area.ping', '');
area.setAttribute('ping', 'p1');
shouldBeEqualToString('area.ping', 'p1');
area.ping = 'p2';
shouldBeEqualToString('area.getAttribute("ping")', 'p2');
area.ping = null;
shouldBeEqualToString('area.getAttribute("ping")', 'null');
shouldBeEqualToString('area.ping', 'null');

</script>
</body>
</html>
